fix(app-services): the library list and the inspector disagreed about MoE models
`GuiModel` — what `GET /api/models` returns — carried no MoE topology, while
`ModelDetailDto` behind `GET /api/models/:id/detail` carried all three fields.
The library list computes *active* parameters from `expertUsedCount /
expertCount`; with both keys absent the formatter fell through to the total.
For a 128-expert model that is wrong by more than an order of magnitude, and
it renders as a plain fact rather than as a fallback, so the inspector and the
list beside it reported different sizes for the same model.

Adds the three fields to `GuiModel` with the same `skip_serializing_if` the
detail DTO uses, so a dense model still omits them rather than growing three
nulls the frontend would need a guard against.

The tests that prove it needed the `Model` builder that already existed inside
the template-caps module, so the model-shaped DTO tests move to their own
`#[path]` sibling and share one fixture — which also keeps `types_tests.rs`
from crossing the 300-line budget as a new file.

Ratchet: types.rs 885 → 905 (three fields, their docs, and the new test
module declaration).
