+++ scripts/rust-complexity-baseline.txt
@@ -17,7 +17,7 @@ crates/gglib-app-services/src/sampling_explain.rs 477
 crates/gglib-app-services/src/servers.rs 965
 crates/gglib-app-services/src/settings.rs 404
 crates/gglib-app-services/src/setup.rs 351
-crates/gglib-app-services/src/types.rs 885
+crates/gglib-app-services/src/types.rs 905
 crates/gglib-axum/src/bootstrap.rs 332
 crates/gglib-axum/src/chat_api.rs 708
 crates/gglib-axum/src/handlers/config/setup.rs 309
